favicon_function/pyproject.toml
[tool.poetry]
name = "favicon-function"
version = "0.0.1"
description = ""
authors = ["Chrrles Paul <charles.paul@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
redis = "^5.0.4"
favicon = "^0.7.0"
fastapi = "^0.111.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
requests-mock = "^1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "favicon_function"
